// Frontend/src/Components/AddExtra/types.ts
export type Division = {
    division: string;
};

export type Shera = {
    shera: string;
};

export type Compliment = {
    compliment: string;
};

export type BirthDetailsInputs = {
    birthPlace: string;
    taluka: string;
    district: string;
    state: string;
    country: string;
};
export type ReligionAndCastes = {
    religion: string;
    caste: string;
    subCaste: string;
};